Update typo
Hi Jeremy!  Enjoyed your post on the differences between hashes in Ruby and objects in Javascript.  I wrote on the same thing and found myself thinking about it a  lot during the course of the week.  Something I've also been thinking about (and I don't think we're far enough along to really grasp this yet) is how inheritances works in Javascript compares to Ruby.  I've also been working on navigating to people's repos, and commenting therein, I think this should point out a typo as well and you should be able to merge the change!